% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{create_model}
\alias{create_model}
\title{Create regression model}
\usage{
create_model(results, perfect_score, score_col, mode = "exp")
}
\arguments{
\item{results}{A dataframe containing archery scores.}

\item{perfect_score}{The maximum number of points possible for the specific archery round}

\item{score_col}{The name of the column in the data frame that contains the scores}

\item{mode}{The type of function ("exp" or "linear") to fit to the results data.}
}
\value{
An R regression model
}
\description{
Fit a linear or exponential model to the trimmed CDF of archery scores.
}
