% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_prob_table.R
\name{generate_prob_table}
\alias{generate_prob_table}
\title{Generate a table of CDF probabilities}
\usage{
generate_prob_table(results, score_col)
}
\arguments{
\item{results}{A data frame containing archery scores}

\item{score_col}{The data fram column containing the scores}
}
\value{
A table of scores and their probabilities according to the
}
\description{
Trim the results and generate a CDF for the score data. Create a new data frame from
the CDF probabilities containing each score and its associated probability. This new data frame
can be used to fit a linear or exponential model
}
