% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_points_table.R
\name{reshape_points_table}
\alias{reshape_points_table}
\title{Reshapes a two-column dataframe of archery scores and Shooter of the Year points.}
\usage{
reshape_points_table(results, cols = 4)
}
\arguments{
\item{results}{A dataframe with a specific (even) number of columns.}

\item{cols}{The number of columns in the reshaped dataframe. It must be an even number (default = 4).}
}
\value{
A new dataframe containing the reshaped dataframe.
}
\description{
Long dataframes are inconvenient to print in a document or web page. This function converts a two-
column dataframe to one with 4, 6, or more columns so it can be printed more compactly.
}
