% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soty.R
\docType{package}
\name{soty}
\alias{soty}
\alias{soty-package}
\title{soty: A package for generating Shooter of the Year points for an archery organization.}
\description{
The soty package provides three categories of important functions:
trim_results.
}
\section{Foo functions}{

The foo functions ...
}

